+++ packages/moonstone/CHANGELOG.md
@@ -5,6 +5,10 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ## [unreleased]
 
+### Added
+
+- `moonstone/Input` and `moonstone/ExpandableInput` support for `<input>` props
+
 ### Fixed
 
 - `moonstone/ExpandableItem` and related expandables to deal with disabled items and the `autoClose`, `lockBottom` and `noLockBottom` props